move connector to a separate file
